second commit

+++ src/main/java/com/entrepreneur/brent_with_sas/repository/SupplierRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SupplierRepo extends JpaRepository<Supplier, Integer> {
+public interface SupplierRepo extends JpaRepository<Supplier, Long> {
 }
